feat: Unify Board and Calendar into HomeView
- Created HomeView to manage BoardView and CalendarView switching
- Consolidated ViewModePicker and AddButton into HomeView toolbar
- Removed redundant HomeToolbar
- Updated MainView to use HomeView
